Fix the decode schedule build, and find why decode cannot be made fast
Chasing "decode should be faster" to the bottom: the reason it is 0.39x is not
the qhead_per_kv=16 rule this report has been citing. That rule is a docstring,
and behind it are four layers.

The first is a build defect. build_decode_schedule/__init__.py never passed
extra_include_paths, so the extension failed on cusparse.h -- torch's
CUDAContextLight.h includes it and a pip CUDA stack ships it only inside the
wheels -- and it pinned -arch=sm_100, which puts a B300 through PTX JIT. Both
were fixed in build_k2q_csr's builder long ago; this one has its own build
configuration and never got them. It now imports those helpers rather than
keeping a second copy. This is why the decode path had never actually been
exercised here: it did not compile.

Past that are three separate guards on qhead_per_kv -- in decode_schedule.py,
in build_decode_schedule.cu, and in interface.py's run path -- none of which is
structural: the code below each derives from qhead_per_kv, and the CuTe kernel
accepts (16, 8, 4, 2, 1).

Relaxing all three reaches the actual wall: "SM100 paged fp8 sparse decode
forward is not implemented yet". The wrapper, the scheduler and the validation
all exist; the forward kernel is a stub. So the fast decode path is not present
in this repo at all, and qhead_per_kv was never the reason -- 16 would not have
worked either.

The three guard relaxations are reverted. They only move the error from "8 is
unsupported" to "not implemented", and leaving them in would imply 16 works.
The right fix is to make those messages point at the missing kernel instead.

probe_decode_kernel.py is left ready: once the forward exists it checks
qhead_per_kv=8 against fp32 over the selected blocks and reports bandwidth
against the selected-KV bytes. Section 17 says what is at stake -- a 0.008 ms
bandwidth floor against 3.1 ms measured.

+++ benchmarks/probe_decode_kernel.py
@@ -61,11 +61,11 @@ def reference(q_b, k_pages, v_pages, sel_b, page_size, scale, group):
     if not blocks:
         return torch.zeros((hq, d), device=q_b.device)
     idx = torch.tensor(blocks, device=q_b.device, dtype=torch.long)
-    k = k_pages[idx].reshape(-1, hkv, d).float()          # [T, Hkv, D]
-    v = v_pages[idx].reshape(-1, hkv, d).float()
+    k = k_pages[idx].reshape(-1, hkv, d).to(torch.float32)
+    v = v_pages[idx].reshape(-1, hkv, d).to(torch.float32)
     k = k.repeat_interleave(group, dim=1)
     v = v.repeat_interleave(group, dim=1)
-    s = torch.einsum("hd,thd->ht", q_b.float(), k) * scale
+    s = torch.einsum("hd,thd->ht", q_b.to(torch.float32), k) * scale
     p = torch.softmax(s, dim=-1)
     return torch.einsum("ht,thd->hd", p, v)
 
@@ -84,7 +84,10 @@ def main() -> int:
 
     from fmha_sm100.cute.interface import SparseDecodePagedAttentionWrapper
 
-    dev, dt = torch.device("cuda"), torch.bfloat16
+    # The paged decode kernel takes an FP8 e4m3 KV cache -- that is a real
+    # format requirement, not a guard: this path exists to read a quantised
+    # cache.  Q is FP8 too.
+    dev, dt = torch.device("cuda"), torch.float8_e4m3fn
     B, S, HQ, HKV, D = args.batch, args.kv_len, args.head_q, args.head_kv, args.dim
     group, scale = HQ // HKV, 1.0 / (D ** 0.5)
     page = args.blk
@@ -137,7 +140,7 @@ def main() -> int:
         ref = reference(q[b], k[b * npage_per_seq:(b + 1) * npage_per_seq],
                         v[b * npage_per_seq:(b + 1) * npage_per_seq],
                         sel[:, b] % npage_per_seq, page, scale, group)
-        c = torch.nn.functional.cosine_similarity(out[b].float(), ref, dim=-1)
+        c = torch.nn.functional.cosine_similarity(out[b].to(torch.float32), ref, dim=-1)
         worst = min(worst, float(c.min()))
         print(f"   请求 {b:>3}: cos 均值={float(c.mean()):.6f} 最差={float(c.min()):.6f}")
